refactor: Float32Array N-API boundary (Phase 3)

+++ include/audio/FFTProcessor.hpp
@@ -5,7 +5,7 @@
 
 namespace fft {
 
-std::vector<std::complex<double>> transform(const std::vector<double> &input);
+std::vector<std::complex<double>> transform(const std::vector<float> &input);
 std::vector<double> inverse(const std::vector<std::complex<double>> &spectrum);
 std::vector<double> magnitude(const std::vector<std::complex<double>> &spectrum);
